Rework details' view models
Pulled common code into a base, abstract view model.
